refacto DisplayItinerary
