--- front/src/applications/osrd/components/Itinerary/DisplayItinerary.js
@@ -1,248 +1,0 @@
-import React from 'react';
-import { useDispatch, useSelector } from 'react-redux';
-import PropTypes from 'prop-types';
-import { RiMapPin2Fill, RiMapPin3Fill, RiMapPin5Fill } from 'react-icons/ri';
-import { GiPathDistance } from 'react-icons/gi';
-import { useTranslation } from 'react-i18next';
-
-import {
-  updateDestination,
-  updateDestinationDate,
-  updateDestinationTime,
-  updateOrigin,
-  updateOriginDate,
-  updateOriginSpeed,
-  updateOriginTime,
-  updateStdcmMode,
-} from 'reducers/osrdconf';
-import { makeEnumBooleans } from 'utils/constants';
-
-import DisplayVias from 'applications/osrd/components/Itinerary/DisplayVias';
-import InputSNCF from 'common/BootstrapSNCF/InputSNCF';
-import { store } from 'Store';
-import { MODES, STDCM_MODES } from '../../consts';
-
-export default function DisplayItinerary(props) {
-  const osrdconf = useSelector((state) => state.osrdconf);
-  const dispatch = useDispatch();
-  const { t } = useTranslation(['osrdconf']);
-  const { zoomToFeaturePoint } = props;
-
-  const { isByOrigin, isByDestination } = makeEnumBooleans(STDCM_MODES, osrdconf.stdcmMode);
-  const { isSimulation, isStdcm } = makeEnumBooleans(MODES, osrdconf.mode);
-
-  return (
-    <div
-      className={
-        osrdconf.origin === undefined &&
-        osrdconf.destination === undefined &&
-        osrdconf.vias.length < 1
-          ? ''
-          : 'osrd-config-anchor'
-      }
-    >
-      <h2 className="d-flex align-items-center mb-0 pl-4">
-        <span className="mr-1 h2 text-success">
-          <RiMapPin2Fill />
-        </span>
-        <span>{t('osrdconf:origin')}</span>
-        <button
-          type="button"
-          data-toggle="modal"
-          data-target="#modalPathJSONDetail"
-          className="btn btn-link"
-        >
-          <small className="ml-1">{osrdconf.pathfindingID}</small>
-        </button>
-      </h2>
-      <div className="mb-3 d-flex align-items-center w-100 osrd-config-place">
-        {osrdconf.origin !== undefined ? (
-          <>
-            <i className="text-success icons-itinerary-bullet mr-2" />
-            <div className="pl-1 hover w-100 d-flex align-items-center">
-              <div
-                onClick={() =>
-                  zoomToFeaturePoint(
-                    osrdconf.origin.clickLngLat,
-                    osrdconf.origin.id,
-                    osrdconf.origin.source
-                  )
-                }
-                role="button"
-                tabIndex={0}
-              >
-                <strong className="mr-1 text-nowrap">
-                  {osrdconf.origin.name ? osrdconf.origin.name : osrdconf.origin.id.split('-')[0]}
-                </strong>
-              </div>
-              <div className="ml-auto d-flex mr-1">
-                {isStdcm && (
-                  <input
-                    type="radio"
-                    id="stdcmMode"
-                    name="stdcmMode"
-                    // className="custom-control-input"
-                    checked={isByOrigin}
-                    onChange={() => dispatch(updateStdcmMode(STDCM_MODES.byOrigin))}
-                  />
-                )}
-                <div className="d-flex">
-                  {isStdcm && (
-                    <input
-                      type="date"
-                      className="form-control form-control-sm"
-                      onChange={(e) => dispatch(updateOriginDate(e.target.value))}
-                      value={osrdconf.originDate}
-                      disabled={isByDestination}
-                    />
-                  )}
-                  <InputSNCF
-                    type="time"
-                    id="osrd-config-time-origin"
-                    onChange={(e) => dispatch(updateOriginTime(e.target.value))}
-                    value={osrdconf.originTime}
-                    sm
-                    noMargin
-                    readonly={isByDestination}
-                  />
-                </div>
-              </div>
-              {isSimulation && (
-                <div className="osrd-config-speed">
-                  <InputSNCF
-                    type="number"
-                    id="osrd-config-speed-origin"
-                    onChange={(e) => dispatch(updateOriginSpeed(e.target.value))}
-                    value={osrdconf.originSpeed}
-                    unit="km/h"
-                    min={0}
-                    max={1000}
-                    sm
-                    noMargin
-                  />
-                </div>
-              )}
-              <button
-                className="btn btn-sm btn-only-icon btn-white"
-                type="button"
-                onClick={() => store.dispatch(updateOrigin(undefined))}
-              >
-                <i className="icons-circle-delete" />
-                <span className="sr-only" aria-hidden="true">
-                  Delete
-                </span>
-              </button>
-            </div>
-          </>
-        ) : (
-          <small className="ml-4">{t('osrdconf:noplacechosen')}</small>
-        )}
-      </div>
-
-      <h2 className="d-flex align-items-center mb-0 ml-4">
-        <span className="mr-1 h2 text-info">
-          <RiMapPin3Fill />
-        </span>
-        <span>{t('osrdconf:vias')}</span>
-        <button
-          className="btn btn-info btn-only-icon btn-sm ml-1"
-          type="button"
-          data-toggle="modal"
-          data-target="#suggeredViasModal"
-        >
-          <GiPathDistance />
-        </button>
-      </h2>
-      <div className="mb-3">
-        {osrdconf.vias.length > 0 ? (
-          <DisplayVias zoomToFeaturePoint={zoomToFeaturePoint} />
-        ) : (
-          <small className="ml-4">{t('osrdconf:noplacechosen')}</small>
-        )}
-      </div>
-
-      <h2 className="d-flex align-items-center mb-0 ml-4">
-        <span className="mr-1 h2 text-warning">
-          <RiMapPin5Fill />
-        </span>
-        <span>{t('osrdconf:destination')}</span>
-      </h2>
-      <div className="mb-3 d-flex align-items-center w-100 osrd-config-place">
-        {osrdconf.destination !== undefined ? (
-          <>
-            <i className="text-warning icons-itinerary-bullet mr-2" />
-            <div className="pl-1 hover w-100 d-flex align-items-center">
-              <div
-                onClick={() =>
-                  zoomToFeaturePoint(
-                    osrdconf.destination.clickLngLat,
-                    osrdconf.destination.id,
-                    osrdconf.destination.source
-                  )
-                }
-                role="button"
-                tabIndex={0}
-                className="flex-grow-1"
-              >
-                <strong className="mr-1 text-nowrap">
-                  {osrdconf.destination.name
-                    ? osrdconf.destination.name
-                    : osrdconf.destination.id.split('-')[0]}
-                </strong>
-              </div>
-
-              {isStdcm && (
-                <div className="ml-auto d-flex mr-1">
-                  <input
-                    type="radio"
-                    id="stdcmMode"
-                    name="stdcmMode"
-                    checked={isByDestination}
-                    onChange={() => dispatch(updateStdcmMode(STDCM_MODES.byDestination))}
-                  />
-
-                  <div className="d-flex">
-                    <input
-                      type="date"
-                      className="form-control form-control-sm"
-                      onChange={(e) => dispatch(updateDestinationDate(e.target.value))}
-                      value={osrdconf.destinationDate}
-                      disabled={isByOrigin}
-                    />
-
-                    <InputSNCF
-                      type="time"
-                      id="osrd-config-time-origin"
-                      onChange={(e) => dispatch(updateDestinationTime(e.target.value))}
-                      value={osrdconf.destinationTime}
-                      sm
-                      noMargin
-                      readonly={isByOrigin}
-                    />
-                  </div>
-                </div>
-              )}
-
-              <button
-                className="btn btn-sm btn-only-icon btn-white"
-                type="button"
-                onClick={() => store.dispatch(updateDestination(undefined))}
-              >
-                <i className="icons-circle-delete" />
-                <span className="sr-only" aria-hidden="true">
-                  Delete
-                </span>
-              </button>
-            </div>
-          </>
-        ) : (
-          <small className="ml-4">{t('osrdconf:noplacechosen')}</small>
-        )}
-      </div>
-    </div>
-  );
-}
-
-DisplayItinerary.propTypes = {
-  zoomToFeaturePoint: PropTypes.func.isRequired,
-};
