+++ front/src/applications/osrd/components/Itinerary/DisplayItinerary/index.ts
@@ -1,0 +1,1 @@
+export { default } from './DisplayItinerary';
